# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_muchlist_moneymagnet_business_pocket_model.NewPocket:
    properties:
      currency:
        example: RP.
        type: string
      editor_id:
        example:
        - 01J4EXF94QDMR5XT9KN527XEP6
        items:
          type: string
        type: array
      icon:
        example: 1
        type: integer
      pocket_name:
        example: dompet utama
        type: string
      watcher_id:
        example:
        - 01J4EXF94QDMR5XT9KN527XEP6
        items:
          type: string
        type: array
    required:
    - pocket_name
    type: object
  github_com_muchlist_moneymagnet_business_pocket_model.PocketResp:
    properties:
      balance:
        example: 50000
        type: integer
      created_at:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      currency:
        example: RP.
        type: string
      editor_id:
        example:
        - 01J4EXF94QDMR5XT9KN527XEP6
        items:
          type: string
        type: array
      icon:
        example: 1
        type: integer
      id:
        example: 01J4EXF94QDMR5XT9KN527XEP8
        type: string
      level:
        example: 1
        type: integer
      owner_id:
        example: 01J4EXF94QDMR5XT9KN527XEP6
        type: string
      pocket_name:
        example: dompet utama
        type: string
      updated_at:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      users:
        items:
          $ref: '#/definitions/github_com_muchlist_moneymagnet_business_pocket_model.PocketUser'
        type: array
      version:
        example: 2
        type: integer
      watcher_id:
        example:
        - 01J4EXF94QDMR5XT9KN527XEP6
        items:
          type: string
        type: array
    type: object
  github_com_muchlist_moneymagnet_business_pocket_model.PocketUpdate:
    properties:
      currency:
        example: RP.
        type: string
      icon:
        example: 1
        type: integer
      pocket_name:
        example: dompet utama
        type: string
    type: object
  github_com_muchlist_moneymagnet_business_pocket_model.PocketUser:
    properties:
      id:
        example: 01J4EXF94QDMR5XT9KN527XEP6
        type: string
      name:
        example: muchlis
        type: string
      role:
        example: owner
        type: string
    type: object
...
  github_com_muchlist_moneymagnet_pkg_data.Metadata:
    properties:
      current_page:
        example: 1
        type: integer
      first_page:
        example: 1
        type: integer
      last_page:
        example: 1
        type: integer
      page_size:
        example: 50
        type: integer
      total_records:
        example: 1
        type: integer
    type: object
  misc.Response500Err:
    properties:
      error:
        example: 'name func: sub func: cause of error'
        type: string
    type: object
  misc.ResponseErr:
    properties:
      error:
        example: example error message
        type: string
      error_field:
        additionalProperties:
          type: string
        example:
          example_field: example_field is a required field
        type: object
    type: object
...
  misc.ResponseSuccess:
    properties:
      data: {}
    type: object
  misc.ResponseSuccessList:
    properties:
      data:
        items: {}
        type: array
      meta_data:
        $ref: '#/definitions/github_com_muchlist_moneymagnet_pkg_data.Metadata'
    type: object
host: localhost
info:
  contact:
    email: whois.muchlis@gmail.com
    name: Muchlis
    url: https://muchlis.dev
  description: this is server for money magnet application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Money Magnet API
  version: "1.0"
paths:
...
  /pockets:
    get:
      consumes:
      - application/json
      description: Find pocket
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page-size
        in: query
        name: page_size
        type: integer
      - description: sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccessList'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_muchlist_moneymagnet_business_pocket_model.PocketResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Find Pocket
      tags:
      - Pocket
    post:
      consumes:
      - application/json
      description: Create Pocket
      parameters:
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_muchlist_moneymagnet_business_pocket_model.NewPocket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/github_com_muchlist_moneymagnet_business_pocket_model.PocketResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Create Pocket
      tags:
      - Pocket
  /pockets/{pocket_id}:
    get:
      consumes:
      - application/json
      description: Get Pocket Detail by ID
      parameters:
      - description: pocket_id
        in: path
        name: pocket_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccessList'
            - properties:
                data:
                  $ref: '#/definitions/github_com_muchlist_moneymagnet_business_pocket_model.PocketResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Get Pocket Detail
      tags:
      - Pocket
    patch:
      consumes:
      - application/json
      description: Update Pocket
      parameters:
      - description: pocket_id
        in: path
        name: pocket_id
        required: true
        type: string
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_muchlist_moneymagnet_business_pocket_model.PocketUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/github_com_muchlist_moneymagnet_business_pocket_model.PocketResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Update Pocket
      tags:
      - Pocket
...
swagger: "2.0"